+++ apps/artemis-web/components/module-state/module-state.spec.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-
 import ModuleState from './module-state';
+import { axe, toHaveNoViolations } from 'jest-axe';
+expect.extend(toHaveNoViolations);
 
 describe('ModuleState', () => {
   it('should render successfully', () => {
@@ -13,5 +14,11 @@ describe('ModuleState', () => {
       </table>
     );
     expect(baseElement).toBeTruthy();
+  });
+
+  it('should have no accessibility violations', async () => {
+    const { baseElement } = render(<ModuleState />);
+    const results = await axe(baseElement);
+    expect(results).toHaveNoViolations();
   });
 });
