+++ apps/artemis-web/components/statistics-table/statistics-table.spec.tsx
@@ -3,6 +3,9 @@ import { render } from '@testing-library/react';
 
 import StatsTable from './statistics-table';
 
+import { axe, toHaveNoViolations } from 'jest-axe';
+expect.extend(toHaveNoViolations);
+
 describe('StatisticsTable', () => {
   it('should render successfully', () => {
     const { baseElement } = render(
@@ -13,5 +16,15 @@ describe('StatisticsTable', () => {
       />
     );
     expect(baseElement).toBeTruthy();
+  });
+
+  it('should have no accessibility violations', async () => {
+    const { baseElement } = render(<StatsTable
+      data={{
+        view_index_all_stats: [{}],
+      }}
+    />);
+    const results = await axe(baseElement);
+    expect(results).toHaveNoViolations();
   });
 });
