+++ apps/artemis-web/pages/_app.tsx
@@ -4,6 +4,22 @@ import { ApolloProvider } from '@apollo/client';
 import { useApollo } from '../libs/graphql';
 import Layout from '../components/layout/layout';
 import TooltipContext from '../context/tooltip-context';
+import ReactDOM from 'react-dom';
+
+if (process.env.NODE_ENV !== 'production' && typeof window !== 'undefined') {
+  /* eslint-disable */
+  const axe = require('react-axe');
+  /* eslint-enable */
+  const config = {
+    rules: [
+      {
+        id: 'radiogroup',
+        enabled: true
+      }
+    ]
+  };
+  axe(React, ReactDOM, 1000, config);
+}
 
 const useStateWithLocalStorage = (localStorageKey) => {
   const [value, setValue] = useState(
