+++ apps/artemis-web/pages/styles.sass
@@ -225,6 +225,14 @@ a:hover
 
 .nav-item > a:hover
   text-decoration: none !important
+  color: white !important
+
+.nav-link
+  color: #dbdbdb !important
+
+.MuiGrid-item > a
+  color: #3f51b5 !important
+  font-weight: bold
 
 .nav-link:hover
   cursor: pointer
