+++ package.json
@@ -108,15 +108,18 @@
     "eslint-plugin-react-hooks": "4.2.0",
     "husky": "^7.0.4",
     "jest": "^27.3.1",
+    "jest-axe": "^5.0.1",
     "jest-fetch-mock": "^3.0.3",
     "msw": "^0.35.0",
     "next-secure-headers": "^2.2.0",
     "node-sass": "^6.0.1",
     "prettier": "2.4.1",
+    "react-axe": "^3.5.4",
     "ts-jest": "^27.0.7",
     "ts-node": "~10.4.0",
     "typescript": "^4.4.4",
-    "url-loader": "^4.1.1"
+    "url-loader": "^4.1.1",
+    "whatwg-fetch": "^3.6.2"
   },
   "husky": {
     "hooks": {
